[#13][#22] Support number rule validation
